Broken code; DO NOT USE.

+++ branches/EnergyPlus/src/GetHPsimInputs.f90
@@ -323,8 +323,8 @@ REAL TWiE   !RS: Indoor Entering or Evaporator Inlet Wetbulb Temperature
 
 !Flow:
 
-!INTEGER, PARAMETER :: r64=KIND(1.0D0)  !RS Comment: Currently needs to be used for integration with Energy+ Code (6/28/12) 
-!REAL(r64), DIMENSION(200) :: TmpNumbers !RS Comment: Currently needs to be used for integration with Energy+ Code (6/28/12)
+!INTEGER, PARAMETER :: r64=KIND(1.0)  !RS Comment: Currently needs to be used for integration with Energy+ Code (6/28/12) 
+!REAL, DIMENSION(200) :: TmpNumbers !RS Comment: Currently needs to be used for integration with Energy+ Code (6/28/12)
 REAL, DIMENSION(200) :: TmpNumbers !RS Comment: Currently needs to be used for integration with Energy+ Code (6/28/12)
 
 CHARACTER(LEN=7),PARAMETER :: FMT_201 = "(10(E))"
